[CORE] .hiverc is loaded twice making create function throw exception #45
**What changes were proposed in this pull request**
If user's `.hiverc` contains sql like 'create temporary function funA as xx.xx', bin/spark-xsql will register it in functionRegistry of local `SparkSession`'s `XSQLSessionCatalog` in the beginning. When some sql causes `SparkSession` restart to run on yarn,  the old `XSQLSessionCatalog` is passed to new started `SparkSession`. It's meaning that funA has existed in functionRegistry. So we should not run `.hiverc` again after SparkSession is restarted.